Rewrite transmuxer to be run more procedurally
