+++ package.json
@@ -18,7 +18,7 @@
     "check": "npm run types && npm run lint",
     "check:demo": "npm run types:demo && npm run lint:demo",
     "check:main": "npm run types:main && npm run lint:main",
-    "check:worker": "npm run types:worker && npm run lint:worker",
+    "check:worker": "npm run types:worker && npm run lint:worker && npm run lint:transmux",
     "clean": "find ./build ! -name 'index.html' ! -name 'favicon.ico' ! -name 'build' ! -name 'style.css' ! -name 'demo.js' ! -name 'logo-white.png' -exec rm -rf {} +",
     "clippy": "cargo clippy",
     "compile:rs": "cargo build --target=wasm32-unknown-unknown && wasm-bindgen --keep-debug --target web ./target/wasm32-unknown-unknown/debug/wasp_hls.wasm --out-dir src/wasm",
@@ -27,10 +27,11 @@
     "fmt": "npm run fmtr && npm run fmtt",
     "fmtr": "cargo fmt",
     "fmtt": "prettier --write . --loglevel warn",
-    "lint": "npm run lint:main && npm run lint:worker && npm run lint:common && npm run lint:demo",
+    "lint": "npm run lint:main && npm run lint:worker && npm run lint:transmux && npm run lint:common && npm run lint:demo",
     "lint:common": "cd src/ts-common && eslint .",
     "lint:demo": "cd demo && eslint src/**/*",
     "lint:main": "cd src/ts-main && eslint .",
+    "lint:transmux": "cd src/ts-worker && eslint .",
     "lint:worker": "cd src/ts-worker && eslint .",
     "move:wasm": "cp src/wasm/wasp_hls_bg.wasm build/wasp_hls_bg.wasm",
     "serve": "http-server ./build",
