--- src/ts-transmux/.eslintrc.json
@@ -1,176 +1,0 @@
-{
-  "env": {
-    "browser": true,
-    "node": true
-  },
-
-  "globals": {
-    "document": false,
-    "escape": false,
-    "navigator": false,
-    "unescape": false,
-    "window": false,
-
-    "ArrayBuffer": true,
-    "DataView": true,
-    "Float32Array": true,
-    "Float64Array": true,
-    "Int16Array": true,
-    "Int32Array": true,
-    "Int8Array": true,
-    "Uint16Array": true,
-    "Uint32Array": true,
-    "Uint8Array": true,
-    "Uint8ClampedArray": true
-  },
-
-  "rules": {
-    "block-scoped-var": 0,
-    "brace-style": [2, "1tbs", { "allowSingleLine": false }],
-    "camelcase": [2, { "properties": "always" }],
-    "comma-dangle": [2, "never"],
-    "comma-spacing": [2, { "before": false, "after": true }],
-    "comma-style": [2, "last"],
-    "complexity": 0,
-    "consistent-return": 0,
-    "consistent-this": 1,
-    "curly": [2, "all"],
-    "default-case": 0,
-    "dot-notation": 2,
-    "eol-last": 2,
-    "eqeqeq": [2, "allow-null"],
-    "func-names": 0,
-    "func-style": [2, "expression"],
-    "guard-for-in": 0,
-    "handle-callback-err": [2, "^(err|error|anySpecificError)$"],
-    "key-spacing": [2, { "beforeColon": false, "afterColon": true }],
-    "keyword-spacing": 2,
-    "max-depth": 0,
-    "max-len": [1, 90, 8, { "ignoreUrls": true }],
-    "max-nested-callbacks": [1, 4],
-    "max-params": 0,
-    "max-statements": 0,
-    "new-cap": [2, { "newIsCap": true, "capIsNew": false }],
-    "new-parens": 2,
-    "newline-after-var": 1,
-    "no-alert": 2,
-    "no-array-constructor": 2,
-    "no-bitwise": 0,
-    "no-caller": 2,
-    "no-catch-shadow": 2,
-    "no-cond-assign": [2, "except-parens"],
-    "no-console": 2,
-    "no-constant-condition": 0,
-    "no-continue": 0,
-    "no-control-regex": 2,
-    "no-debugger": 2,
-    "no-delete-var": 2,
-    "no-div-regex": 0,
-    "no-dupe-args": 2,
-    "no-dupe-keys": 2,
-    "no-duplicate-case": 2,
-    "no-else-return": 2,
-    "no-empty": 2,
-    "no-empty-character-class": 2,
-    "no-labels": 2,
-    "no-eq-null": 0,
-    "no-eval": 2,
-    "no-ex-assign": 2,
-    "no-extend-native": 2,
-    "no-extra-bind": 2,
-    "no-extra-boolean-cast": 2,
-    "no-extra-parens": 0,
-    "no-extra-semi": 2,
-    "no-fallthrough": 2,
-    "no-floating-decimal": 2,
-    "no-func-assign": 2,
-    "no-implied-eval": 2,
-    "no-inner-declarations": [2, "functions"],
-    "no-invalid-regexp": 2,
-    "no-irregular-whitespace": 2,
-    "no-iterator": 2,
-    "no-label-var": 2,
-    "no-labels": 2,
-    "no-lone-blocks": 2,
-    "no-lonely-if": 2,
-    "no-loop-func": 1,
-    "no-mixed-requires": [0, false],
-    "no-mixed-spaces-and-tabs": [2, false],
-    "no-multi-spaces": 1,
-    "no-multi-str": 2,
-    "no-multiple-empty-lines": [2, { "max": 2 }],
-    "no-native-reassign": 2,
-    "no-negated-in-lhs": 2,
-    "no-nested-ternary": 1,
-    "no-new": 2,
-    "no-new-func": 2,
-    "no-new-object": 2,
-    "no-new-require": 2,
-    "no-new-wrappers": 2,
-    "no-obj-calls": 2,
-    "no-octal": 2,
-    "no-octal-escape": 2,
-    "no-path-concat": 2,
-    "no-param-reassign": 0,
-    "no-plusplus": 0,
-    "no-process-env": 0,
-    "no-process-exit": 0,
-    "no-proto": 2,
-    "no-redeclare": 2,
-    "no-regex-spaces": 2,
-    "no-restricted-modules": 0,
-    "no-return-assign": 2,
-    "no-script-url": 2,
-    "no-self-compare": 1,
-    "no-sequences": 2,
-    "no-shadow": 1,
-    "no-shadow-restricted-names": 2,
-    "no-spaced-func": 2,
-    "no-sparse-arrays": 2,
-    "no-sync": 0,
-    "no-ternary": 0,
-    "no-throw-literal": 2,
-    "no-trailing-spaces": 2,
-    "no-undef": 2,
-    "no-undef-init": 2,
-    "no-undefined": 0,
-    "no-underscore-dangle": 0,
-    "no-unreachable": 2,
-    "no-unused-expressions": 2,
-    "no-unused-vars": [2, { "vars": "all", "args": "none" }],
-    "no-use-before-define": 2,
-    "no-void": 0,
-    "no-warning-comments": [
-      1,
-      { "terms": ["todo", "fixme", "xxx"], "location": "start" }
-    ],
-    "no-with": 2,
-    "operator-assignment": [0, "always"],
-    "operator-linebreak": [2, "after"],
-    "padded-blocks": 0,
-    "quote-props": [2, "as-needed"],
-    "quotes": [2, "single", "avoid-escape"],
-    "radix": 2,
-    "semi": [2, "always"],
-    "semi-spacing": [2, { "before": false, "after": true }],
-    "sort-vars": 0,
-    "space-before-blocks": [2, "always"],
-    "space-before-function-paren": [
-      2,
-      { "anonymous": "never", "named": "never" }
-    ],
-    "space-in-brackets": 0,
-    "space-in-parens": [2, "never"],
-    "space-infix-ops": 2,
-    "space-unary-ops": [2, { "words": true, "nonwords": false }],
-    "spaced-comment": [2, "always"],
-    "strict": 0,
-    "use-isnan": 2,
-    "valid-jsdoc": 0,
-    "valid-typeof": 2,
-    "vars-on-top": 1,
-    "wrap-iife": [2, "outside"],
-    "wrap-regex": 2,
-    "yoda": [2, "never", { "exceptRange": true }]
-  }
-}
