+++ src/ts-transmux/index.ts
@@ -1,2 +1,2 @@
-import Transmuxer from "./src/transmuxer";
+import Transmuxer from "./src/Transmuxer";
 export default Transmuxer;
