+++ src/ts-transmux/src/constants.ts
@@ -1,18 +1,3 @@
-const AUDIO_PROPERTIES = [
-  "audioobjecttype",
-  "channelcount",
-  "samplerate",
-  "samplingfrequencyindex",
-  "samplesize",
-] as const;
-
-const VIDEO_PROPERTIES = [
-  "width",
-  "height",
-  "profileIdc",
-  "levelIdc",
-  "profileCompatibility",
-  "sarRatio",
-];
-
-export { AUDIO_PROPERTIES, VIDEO_PROPERTIES };
+export const H264_STREAM_TYPE = 0x1b;
+export const ADTS_STREAM_TYPE = 0x0f;
+export const METADATA_STREAM_TYPE = 0x15;
