+++ src/ts-worker/globals.ts
@@ -2,6 +2,7 @@ import { numberIdGenerator } from "../ts-common/idGenerator";
 import logger from "../ts-common/logger";
 import type QueuedSourceBuffer from "../ts-common/QueuedSourceBuffer";
 import type { SourceBufferId, WaspHlsPlayerConfig } from "../ts-common/types";
+import type Transmuxer from "../ts-transmux";
 import { Dispatcher, type InitOutput, type MediaType } from "../wasm/wasp_hls";
 
 export interface WorkerInitializationOptions {
@@ -115,7 +116,7 @@ export interface SourceBufferInstanceInfo<HasMseInWorker extends boolean> {
   mediaType: MediaType;
   lastInitTimescale: number | undefined;
   sourceBuffer: HasMseInWorker extends true ? QueuedSourceBuffer : null;
-  transmuxer: null | ((inputSegment: Uint8Array) => Uint8Array | null);
+  transmuxer: null | Transmuxer;
 }
 
 export interface WorkerMediaSourceInstanceInfo {
